Update index.html to have all the content
I created containers for each of the sections and commented each
section and what flex-direction I will need to use for reference. I
sorted the information this way because I was thinking of everything as
boxes and created the appropriate containers so that I can style them
the way I think it is meant to be styled and efficiently too.